Move BoundsSafetyArgs.{cpp,h} from clangDriver to clangOptions
This is needed to fix link errors on the bots in `ninja ClangIncludeCleanerTests` (and possibly elsewhere) after merging https://github.com/llvm/llvm-project/pull/165277 .

According to that PR's description, "clangFrontend must no longer depend on clangDriver".

In our downstream build, we had such a dependency from clang/lib/Frontend/CompilerInvocation.cpp to clang/lib/Driver/BoundsSafetyArgs.cpp for `ParseBoundsSafetyNewChecksMaskFromArgs`.

I'm moving BoundsSafetyArgs to Options to break that dependency. Maybe it would be better in Frontend, I'm not sure. For now this unbreaks the build and I'll ask code owners to review after the US holiday.

(cherry picked from commit f9c24de9194c6088bd1a554a568b1bc2d87299c2)
